# Conflicts:
#	Windows/InedoExtension/InedoExtension.csproj
#	Windows/InedoExtension/PowerShell/PSUtil.cs
#	Windows/InedoExtension/packages.config

+++ Windows/InedoExtension/Operations/Registry/EnsureRegistryKeyOperation.cs
@@ -38,12 +38,15 @@ namespace Inedo.Extensions.Windows.Operations.Registry
             {
                 if (key == null)
                 {
+                    this.LogInformation(this.Template.GetDisplayPath() + " does not exist.");
                     config.Exists = false;
                 }
                 else
                 {
                     config.Exists = true;
                     config.DefaultValue = key.GetValue(null)?.ToString();
+
+                    this.LogInformation(this.Template.GetDisplayPath() + " exists.");
                 }
             }
 
@@ -52,19 +55,38 @@ namespace Inedo.Extensions.Windows.Operations.Registry
 
         protected override Task RemoteConfigureAsync(IRemoteOperationExecutionContext context)
         {
+            this.LogDebug($"Configuring {this.Template.GetDisplayPath()}...");
+
             using (var baseKey = RegistryKey.OpenBaseKey(this.Template.Hive, RegistryView.Default))
             {
                 if (this.Template.Exists)
                 {
-                    using (var key = baseKey.CreateSubKey(this.Template.Key))
+                    this.LogInformation($"Ensuring that {this.Template.GetDisplayPath()} exists...");
+
+                    using (var key = createOrOpen())
                     {
+                        this.LogInformation(this.Template.GetDisplayPath() + " created.");
                         if (!string.IsNullOrWhiteSpace(this.Template.DefaultValue))
-                            key.SetValue(null, this.Template.DefaultValue);
+                        {
+                            this.LogDebug($"Setting default value to {this.Template.DefaultValue}...");
+                            if (!context.Simulation)
+                                key.SetValue(null, this.Template.DefaultValue);
+                        }
                     }
                 }
                 else
                 {
-                    baseKey.DeleteSubKeyTree(this.Template.Key, false);
+                    this.LogInformation($"Deleting {this.Template.GetDisplayPath()}...");
+                    if (!context.Simulation)
+                        baseKey.DeleteSubKeyTree(this.Template.Key, false);
+                }
+
+                RegistryKey createOrOpen()
+                {
+                    if (context.Simulation)
+                        return baseKey.OpenSubKey(this.Template.Key);
+                    else
+                        return baseKey.CreateSubKey(this.Template.Key);
                 }
             }
